Rename and delete examples.
