+++ huntflow_api_client/models/response/applicant_on_vacancy_status.py
